+++ class/Cryptex.php
@@ -1,13 +1,13 @@
 <?php
 /**
     Cryptex Class
-    Version: 5.1
+    Version: 5.2
     Author: Andi Dittrich
     Author URI: http://andidittrich.de
     Plugin URI: http://andidittrich.de/go/cryptex
     License: MIT X11-License
     
-    Copyright (c) 2010-2015, Andi Dittrich
+    Copyright (c) 2010-2016, Andi Dittrich
 
     Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
     
@@ -27,7 +27,7 @@ class Cryptex{
     
     // get singelton instance
     public static function getInstance(){
-        // check if singelton instance is avaible
+        // check if singleton instance is available
         if (self::$__instance==null){
             // create new instance if not
             self::$__instance = new self();
@@ -43,7 +43,7 @@ class Cryptex{
             'email-replacement-dot' => '.',
             'embed-css' => true,
             'embed-js' => true,
-            'font-file' => 'Arial.ttf',
+            'font-file' => 'LiberationSans-Regular.ttf',
             'font-family' => 'inherit',
             'font-size' => '12px',
             'line-height' => '0',